Autogenerate shortened on persist if empty
