+++ config/services.yaml
@@ -29,3 +29,8 @@ services:
 
     # add more service definitions when explicit configuration is needed
     # please note that last definitions always *replace* previous ones
+    App\EntityListener\UrlEntityListener:
+        tags:
+            - name: doctrine.orm.entity_listener
+              event: prePersist
+              entity: App\Entity\Url
